Add generic PathMethod class
